Split off the in-memory store to a separate source file.
