+++ lib/staticGzip.js
@@ -13,6 +13,7 @@ var fs = require('fs'),
 		path = require('path'),
 		mime = require('mime'),
 		compress = require('compress'),
+		gzippoCache = require('./memoryStore'),
 		staticSend;
 try {
 	staticSend = require('connect').static.send;
@@ -34,23 +35,6 @@ var removeContentHeaders = function(res){
 		}
 	});
 };
-
-/**
- * gzipped in-memory cache.
- */
-
-var gzippoCache = (function() {
-	var cache = {};
-	this.get = function(pathname) {
-		return cache[pathname];
-	};
-	this.set = function(pathname, cacheObj) {
-		var oldCache = typeof(cache[pathname]) == "object" ? cache[pathname] : undefined;
-		cache[pathname] = cacheObj;
-		return oldCache;
-	};
-	return this;
-})();
 
 /**
  * gzip file.
